Add window.TikZJaxNoAutostart
